move logic to a separate package
